Fix dashboard urls
