+++ pgml-dashboard/src/api/mod.rs
@@ -2,6 +2,7 @@ use rocket::route::Route;
 
 pub mod chatbot;
 pub mod cms;
+pub mod deployment;
 
 pub fn routes() -> Vec<Route> {
     let mut routes = Vec::new();
