+++ pgml-dashboard/src/guards.rs
@@ -1,5 +1,6 @@
 use crate::components::sections::footers::marketing_footer::MarketingFooter;
 use crate::templates::components::{StaticNav, StaticNavLink};
+use crate::utils::urls;
 use once_cell::sync::OnceCell;
 use rocket::http::Status;
 use rocket::request::{self, FromRequest, Request};
@@ -50,25 +51,13 @@ impl Cluster {
                 },
                 product_left_nav: StaticNav {
                     links: vec![
-                        StaticNavLink::new("Notebooks".to_string(), "/dashboard?tab=Notebooks".to_string())
-                            .icon("add_notes")
-                            .active(
-                                uri.is_some()
-                                    && (uri.clone().unwrap().starts_with("/dashboard?tab=Notebook")
-                                        || uri.clone().unwrap() == "/dashboard"),
-                            ),
-                        StaticNavLink::new("Projects".to_string(), "/dashboard?tab=Projects".to_string())
-                            .icon("library_add")
-                            .active(uri.is_some() && uri.clone().unwrap().starts_with("/dashboard?tab=Project")),
-                        StaticNavLink::new("Models".to_string(), "/dashboard?tab=Models".to_string())
-                            .icon("space_dashboard")
-                            .active(uri.is_some() && uri.clone().unwrap().starts_with("/dashboard?tab=Model")),
-                        StaticNavLink::new("Snapshots".to_string(), "/dashboard?tab=Snapshots".to_string())
-                            .icon("filter_center_focus")
-                            .active(uri.is_some() && uri.clone().unwrap().starts_with("/dashboard?tab=Snapshot")),
-                        StaticNavLink::new("Upload data".to_string(), "/dashboard?tab=Upload_Data".to_string())
-                            .icon("upload")
-                            .active(uri.is_some() && uri.clone().unwrap().starts_with("/dashboard?tab=Upload_Data")),
+                        StaticNavLink::new("Notebooks".to_string(), urls::deployment_notebooks())
+                            .icon("format_list_bulleted_add"),
+                        StaticNavLink::new("Projects".to_string(), urls::deployment_projects()).icon("library_add"),
+                        StaticNavLink::new("Models".to_string(), urls::deployment_models()).icon("grid_view"),
+                        StaticNavLink::new("Snapshots".to_string(), urls::deployment_snapshots())
+                            .icon("filter_center_focus"),
+                        StaticNavLink::new("Upload data".to_string(), urls::deployment_uploader()).icon("upload"),
                     ],
                 },
                 marketing_footer: MarketingFooter::new().render_once().unwrap(),
