+++ pgml-dashboard/templates/content/dashboard/panels/model.html
@@ -1,3 +1,4 @@
+<% use crate::utils::urls; %>
 <turbo-frame id="model">
   <div class="notebook">
     <section>
@@ -12,10 +13,10 @@
     
       <dl>
         <dt>Project</dt>
-        <dd><a href="/dashboard?tab=Project&id=<%= project.id %>" data-turbo="false"><%= project.name %></a></dd>
+        <dd><a href="<%- urls::deployment_project_by_id(project.id) %>" data-turbo="false"><%= project.name %></a></dd>
         <% if let Some(snapshot) = snapshot { %>
         <dt>Snapshot</dt>
-        <dd><a href="/dashboard?tab=Snapshot&id=<%= snapshot.id %>" data-turbo="false"><%= snapshot.relation_name %></a></dd>
+        <dd><a href="<%- urls::deployment_snapshot_by_id(snapshot.id) %>" data-turbo="false"><%= snapshot.relation_name %></a></dd>
         <% } %>
         <dt>Created</dt>
         <dd><time datetime="<% model.created_at; %>"><% model.created_at; %></time></dd>
