+++ pgml-dashboard/templates/content/dashboard/tabs/model_tab.html
@@ -1,2 +1,2 @@
-<turbo-frame id="model" src="/dashboard/models/<%= model_id %>">
+<turbo-frame id="model" src="/dashboard/models_turboframe/<%= model_id %>">
 </turbo-frame>
