+++ pgml-dashboard/templates/content/dashboard/tabs/models_tab.html
@@ -1,4 +1,4 @@
-<turbo-frame id="models" src="/dashboard/models" target="_top">
+<turbo-frame id="models" src="/dashboard/models_turboframe" target="_top">
     <div class="d-flex justify-content-center align-items-center py-5">Loading models, one moment...</div>
 </turbo-frame>
