+++ pgml-dashboard/templates/content/dashboard/tabs/notebook_tab.html
@@ -1,4 +1,4 @@
 
-<turbo-frame id="notebook" src="/dashboard/notebooks/<%= id %>">
+<turbo-frame id="notebook" src="/dashboard/notebooks_turboframe/<%= id %>">
     <div class="d-flex justify-content-center align-items-center py-5">Loading notebook, one moment...</div>
 </turbo-frame>
