+++ pgml-dashboard/templates/content/dashboard/tabs/notebooks_tab.html
@@ -1,4 +1,4 @@
 
-<turbo-frame id="notebooks" src="/dashboard/notebooks">
+<turbo-frame id="notebooks" src="/dashboard/notebooks_turboframe">
   <div class="d-flex justify-content-center align-items-center py-5">Loading notebooks, one moment...</div>
 </turbo-frame>
