+++ pgml-dashboard/templates/content/dashboard/tabs/project_tab.html
@@ -1,4 +1,4 @@
 
-<turbo-frame id="project" src="/dashboard/projects/<%= project_id %>">
+<turbo-frame id="project" src="/dashboard/projects_turboframe/<%= project_id %>">
   <div class="d-flex justify-content-center align-items-center py-5">Loading project, one moment...</div>
 </turbo-frame>
