+++ pgml-dashboard/templates/content/dashboard/tabs/snapshot_tab.html
@@ -1,2 +1,2 @@
-<turbo-frame id="snapshot" src="/dashboard/snapshots/<%= snapshot_id %>">
+<turbo-frame id="snapshot" src="/dashboard/snapshots_turboframe/<%= snapshot_id %>">
 </turbo-frame>
