+++ pgml-dashboard/templates/content/dashboard/tabs/snapshots_tab.html
@@ -1,4 +1,4 @@
-<turbo-frame id="snapshots" src="/dashboard/snapshots">
+<turbo-frame id="snapshots" src="/dashboard/snapshots_turboframe">
     <div class="d-flex justify-content-center align-items-center py-5">Loading snapshots, one moment...</div>
 </turbo-frame>
