+++ pgml-dashboard/templates/content/dashboard/tabs/uploader_tab.html
@@ -1,11 +1,11 @@
 <%
     let source = match table_name {
-        Some(name) => format!("/dashboard/uploader/done?table_name={}", name), 
+        Some(name) => format!("/dashboard/uploader_turboframe/done?table_name={}", name), 
         None => "".to_string(),
     };
 %>
 
-<turbo-frame id="upload" src="/dashboard/uploader">
+<turbo-frame id="upload" src="/dashboard/uploader_turboframe">
     <div class="d-flex justify-content-center align-items-center py-5">Loading snapshots, one moment...</div>
 </turbo-frame>
 <turbo-frame id="uploaded" src="<%= source %>">
